# Class transformer optimization