### What?

This PR optimizes some Forge class transformers, greatly reducing startup time in a dev environment.
 - `DeobfuscationTransformer`: This is run on external libraries, which should never contain references to anything in the MC source, they can be skipped. Additionally, in a dev environment, vanilla class names are already all deobfuscated, so these can be skipped too.
 - `TerminalTransformer`: Minor optimization; currently serializes classes even if they were not modified at all. It has been changed here to skip serialization if nothing changed.

### Performance
This can have a large impact on performance on slower CPUs, as can be seen here in profiler results during startup:
![Before](https://i.imgur.com/ovaSJgE.png)
->
![After](https://i.imgur.com/L6eKbjU.png)