@liach 2018-09-24
i suggest sorting the array and looking up strings with binary search