@gegy1000 2018-09-24
@liach Could do, though I'm not sure if it's worth it. 8s->100ms is not a huge difference from, for example, 8s->10ms with a more optimized lookup.